pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "relai"
requires-python = ">=3.9"
authors = [
    {name = "RELAI", email = "support@relai.ai"},
]
description = "An SDK for building reliable AI agents"
readme = "README.md"
license = {file = "LICENSE.md" }
classifiers = [
  "License :: Other/Proprietary License",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
]

dependencies = [
    "pydantic>=2.11.5",
    "numpy>=1.26.4",
    "aiohttp[speedups]>=3.12.15",
    "httpx>=0.28.1",
    "openai-agents[litellm]>=0.2.10",
    "opentelemetry-instrumentation>=0.58b0",
    "openinference-instrumentation>=0.1.38",
]
version = "0.3.0"

[tool.setuptools.packages.find]
where = ["."]
include = ["relai*"]

[dependency-groups]
dev = [
    "build>=1.3.0",
    "google-genai>=1.45.0",
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.6.14",
    "mkdocstrings[python]>=0.29.1",
    "openai==1.102.0",
    "openai-agents>=0.2.10",
    "openinference-instrumentation-openai-agents>=1.3.0",
    "pytest>=8.3.5",
    "pytest-asyncio>=1.1.0",
    "pytest-cov>=6.3.0",
    "pytest-mock>=3.15.0",
    "responses>=0.25.7",
    "ruff>=0.11.11",
]
